ai-client/pc/requirements.txt
requests
python-dotenv
sounddevice
scipy
pydub